requirement.txt
streamlit
torch
torchvision
transformers
scikit-learn
tqdm
numpy
pandas
pillow
requests 
